Move the range argument providers to the params package
