+++ src/main/java/org/junitpioneer/jupiter/params/LongRange.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junitpioneer.jupiter;
+package org.junitpioneer.jupiter.params;
 
 public class LongRange extends Range<Long> {
 	public LongRange(LongRangeSource source) {
