+++ src/main/java/org/junitpioneer/jupiter/params/RangeSourceProvider.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junitpioneer.jupiter;
+package org.junitpioneer.jupiter.params;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
