+++ src/main/java/org/junitpioneer/jupiter/params/package-info.java
@@ -1,0 +1,15 @@
+/**
+ * Argument providers for a range of numbers.
+ *
+ * <p>Check out the following types for details:
+ * <ul>
+ *     <li>{@link org.junitpioneer.jupiter.params.ByteRange}</li>
+ *     <li>{@link org.junitpioneer.jupiter.params.DoubleRange}</li>
+ *     <li>{@link org.junitpioneer.jupiter.params.FloatRange}</li>
+ *     <li>{@link org.junitpioneer.jupiter.params.IntRange}</li>
+ *     <li>{@link org.junitpioneer.jupiter.params.LongRange}</li>
+ *     <li>{@link org.junitpioneer.jupiter.params.ShortRange}</li>
+ * </ul>
+ */
+
+package org.junitpioneer.jupiter.params;
